TST: filter ComplexWarning from fblas tests.
(backport of r7036)

+++ scipy/lib/blas/tests/test_fblas.py
@@ -89,7 +89,7 @@ class BaseAxpy(object):
         assert_(0)
 
     def test_y_bad_size(self):
-        x = arange(12.,dtype=complex64)
+        x = arange(12.,dtype=self.dtype)
         y = zeros(6,x.dtype)
         try:
             self.blas_func(x,y,n=3,incy=5)
@@ -217,7 +217,7 @@ class BaseCopy(object):
         assert_(0)
 
     def test_y_bad_size(self):
-        x = arange(12.,dtype=complex64)
+        x = arange(12.,dtype=self.dtype)
         y = zeros(6,x.dtype)
         try:
             self.blas_func(x,y,n=3,incy=5)
@@ -310,7 +310,7 @@ class BaseSwap(object):
         assert_(0)
 
     def test_y_bad_size(self):
-        x = arange(12.,dtype=complex64)
+        x = arange(12.,dtype=self.dtype)
         y = zeros(6,x.dtype)
         try:
             self.blas_func(x,y,n=3,incy=5)
